resolve conversation

+++ src/game.h
@@ -62,7 +62,7 @@ static constexpr int32_t RANGE_WRAP_ITEM_INTERVAL = 400;
 static constexpr int32_t RANGE_REQUEST_TRADE_INTERVAL = 400;
 
 static constexpr uint8_t ITEM_STACK_SIZE = 100;
-static constexpr uint8_t MAX_STACKPOS = 10;
+static constexpr int32_t MAX_STACKPOS = 10;
 
 /**
  * Main Game class.
